+++ controller/dataplane/owned_resources.go
@@ -13,8 +13,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	policyv1 "k8s.io/api/policy/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/selection"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
@@ -199,14 +197,7 @@ func ensureAdminServiceForDataPlane(
 	additionalServiceLabels client.MatchingLabels,
 	opts ...k8sresources.ServiceOpt,
 ) (res op.Result, svc *corev1.Service, err error) {
-	// TODO: https://github.com/Kong/gateway-operator/issues/156.
-	// Use only new labels after several minor version of soak time.
-
-	// Below we list both the Services with the new labels and the legacy labels
-	// in order to support upgrades from older versions of the operator and perform
-	// the reduction of the Services using the older labels.
-
-	// Get the Services for the DataPlane using new labels.
+	// Get the Services for the DataPlane by label.
 	matchingLabels := k8sresources.GetManagedLabelForOwner(dataPlane)
 	matchingLabels[consts.DataPlaneServiceTypeLabel] = string(consts.DataPlaneAdminServiceLabelValue)
 	for k, v := range additionalServiceLabels {
@@ -223,31 +214,6 @@ func ensureAdminServiceForDataPlane(
 	if err != nil {
 		return op.Noop, nil, fmt.Errorf("failed listing Services for DataPlane %s/%s: %w", dataPlane.Namespace, dataPlane.Name, err)
 	}
-
-	// Get the Services for the DataPlane using legacy labels.
-	reqLegacyLabels, err := k8sresources.GetManagedLabelRequirementsForOwnerLegacy(dataPlane)
-	if err != nil {
-		return op.Noop, nil, err
-	}
-	reqLegacyServiceType, err := labels.NewRequirement(
-		consts.DataPlaneServiceTypeLabelLegacy, selection.Equals, []string{string(consts.DataPlaneAdminServiceLabelValue)},
-	)
-	if err != nil {
-		return op.Noop, nil, err
-	}
-	servicesLegacy, err := k8sutils.ListServicesForOwner(
-		ctx,
-		cl,
-		dataPlane.Namespace,
-		dataPlane.UID,
-		&client.ListOptions{
-			LabelSelector: labels.NewSelector().Add(*reqLegacyServiceType).Add(reqLegacyLabels...),
-		},
-	)
-	if err != nil {
-		return op.Noop, nil, fmt.Errorf("failed listing Services for DataPlane %s/%s: %w", dataPlane.Namespace, dataPlane.Name, err)
-	}
-	services = append(services, servicesLegacy...)
 
 	count := len(services)
 	if count > 1 {
@@ -310,14 +276,7 @@ func ensureIngressServiceForDataPlane(
 	additionalServiceLabels client.MatchingLabels,
 	opts ...k8sresources.ServiceOpt,
 ) (op.Result, *corev1.Service, error) {
-	// TODO: https://github.com/Kong/gateway-operator/issues/156.
-	// Use only new labels after several minor version of soak time.
-
-	// Below we list both the Services with the new labels and the legacy labels
-	// in order to support upgrades from older versions of the operator and perform
-	// the reduction of the Services using the older labels.
-
-	// Get the Services for the DataPlane using new labels.
+	// Get the Services for the DataPlane by label.
 	matchingLabels := k8sresources.GetManagedLabelForOwner(dataPlane)
 	matchingLabels[consts.DataPlaneServiceTypeLabel] = string(consts.DataPlaneIngressServiceLabelValue)
 	for k, v := range additionalServiceLabels {
@@ -334,31 +293,6 @@ func ensureIngressServiceForDataPlane(
 	if err != nil {
 		return op.Noop, nil, fmt.Errorf("failed listing Services for DataPlane %s/%s: %w", dataPlane.Namespace, dataPlane.Name, err)
 	}
-
-	// Get the Services for the DataPlane using legacy labels.
-	reqLegacyLabels, err := k8sresources.GetManagedLabelRequirementsForOwnerLegacy(dataPlane)
-	if err != nil {
-		return op.Noop, nil, err
-	}
-	reqLegacyServiceType, err := labels.NewRequirement(
-		consts.DataPlaneServiceTypeLabelLegacy, selection.Equals, []string{string(consts.DataPlaneProxyServiceLabelValueLegacy)},
-	)
-	if err != nil {
-		return op.Noop, nil, err
-	}
-	servicesLegacy, err := k8sutils.ListServicesForOwner(
-		ctx,
-		cl,
-		dataPlane.Namespace,
-		dataPlane.UID,
-		&client.ListOptions{
-			LabelSelector: labels.NewSelector().Add(*reqLegacyServiceType).Add(reqLegacyLabels...),
-		},
-	)
-	if err != nil {
-		return op.Noop, nil, fmt.Errorf("failed listing Services for DataPlane %s/%s: %w", dataPlane.Namespace, dataPlane.Name, err)
-	}
-	services = append(services, servicesLegacy...)
 
 	count := len(services)
 	if count > 1 {